refactor: 서비스로직, 핵심기능로직 분리
