+++ src/domain/LottoCompare.js
