+++ src/domain/LottosGenerator.js
