+++ src/domain/LottosResults.js
