+++ src/domain/LottosReturns.js
